Fix resizing issues
When the real Twitter API is on the page and working there was an issue where I was erroneously resizing resizing the wrong iframe. Fixing this by making sure we send the actual dom id of the iframe we make, only resizing iframes we make, and only listening for resizing events.